Combine EthProvider and EthProviderPw into a single instance that autodetects the lock used.
