+++ src/providers/eth-provider.ts
@@ -1,12 +1,18 @@
+import PWCore, { ChainID } from '../core';
+import { Address, AddressType, LockType, LockTypeOmniPw } from '../models';
 import { Provider, Platform } from './provider';
-import { Address, AddressType } from '../models';
 import { Message } from '../signers';
 import ENS from 'ethereum-ens';
 
 export class EthProvider extends Provider {
+  lockType: LockTypeOmniPw | null;
   onAddressChanged: (newAddress: Address) => void;
-  constructor(onAddressChanged?: (newAddress: Address) => void) {
+  constructor(
+    onAddressChanged?: (newAddress: Address) => void,
+    lockType: LockTypeOmniPw | null = null
+  ) {
     super(Platform.eth);
+    this.lockType = lockType;
     this.onAddressChanged = onAddressChanged;
   }
   async init(): Promise<Provider> {
@@ -15,11 +21,21 @@ export class EthProvider extends Provider {
       const accounts = await window.ethereum.request({
         method: 'eth_requestAccounts',
       });
-      this.address = new Address(accounts[0], AddressType.eth);
+      this.address = new Address(
+        accounts[0],
+        AddressType.eth,
+        undefined,
+        this.lockType
+      );
 
       if (!!window.ethereum.on) {
         window.ethereum.on('accountsChanged', (newAccounts: string[]) => {
-          this.address = new Address(newAccounts[0], AddressType.eth);
+          this.address = new Address(
+            newAccounts[0],
+            AddressType.eth,
+            undefined,
+            this.lockType
+          );
           if (!!this.onAddressChanged) {
             this.onAddressChanged(this.address);
           }
@@ -37,7 +53,12 @@ export class EthProvider extends Provider {
           resolve(result);
         });
       });
-      this.address = new Address(accounts[0], AddressType.eth);
+      this.address = new Address(
+        accounts[0],
+        AddressType.eth,
+        undefined,
+        this.lockType
+      );
 
       return this;
     } else {
@@ -55,7 +76,7 @@ export class EthProvider extends Provider {
     }
   }
 
-  protected handleResult(result): string {
+  protected handleResultOmni(result): string {
     let v = Number.parseInt(result.slice(-2), 16);
     if (v >= 27) v -= 27;
     result =
@@ -66,15 +87,34 @@ export class EthProvider extends Provider {
     return result;
   }
 
+  protected handleResultPw = (result): string => {
+    let v = Number.parseInt(result.slice(-2), 16);
+    if (v >= 27) v -= 27;
+    result =
+      '0x' +
+      (PWCore.chainId === ChainID.ckb
+        ? ''
+        : this.platform.toString(16).padStart(2, '0')) + // Testnet release requires an extra byte for the platform.
+      result.slice(2, -2) +
+      v.toString(16).padStart(2, '0');
+    return result;
+  };
+
   async sign(message: Message): Promise<string> {
     return new Promise((resolve, reject) => {
+      // Determine handler based on the lock type.
+      const handleResult =
+        message.lock.identifyLockType() === LockType.pw
+          ? this.handleResultPw
+          : this.handleResultOmni;
+
       const from = this.address.addressString;
 
       if (typeof window.ethereum !== 'undefined') {
         window.ethereum
           .request({ method: 'personal_sign', params: [from, message.message] })
           .then((result) => {
-            resolve(this.handleResult(result));
+            resolve(handleResult(result));
           });
       } else if (!!window.web3) {
         window.web3.currentProvider.sendAsync(
@@ -86,7 +126,7 @@ export class EthProvider extends Provider {
             if (result.error) {
               reject(result.error);
             }
-            resolve(this.handleResult(result.result));
+            resolve(handleResult(result.result));
           }
         );
       } else {
