+++ src/providers/index.ts
@@ -1,7 +1,6 @@
 export * from './provider';
 export * from './dummy-provider';
 export * from './eth-provider';
-export * from './eth-provider-pw';
 export * from './eos-provider';
 export * from './null-provider';
 export * from './tron-provider';
